refactor: abstraction for altpayments
