+++ react/lib/util/socket.ts
@@ -1,6 +1,6 @@
 import { Socket } from 'socket.io-client';
 
-import { BroadcastTxData, SideshiftCoin, SideshiftError, SideshiftPair, SideshiftShift } from './types';
+import { BroadcastTxData } from './types';
 
 export const txsListener = (socket: Socket, setNewTxs: Function): void => {
   socket.on('incoming-txs', (broadcastedTxData: BroadcastTxData) => {
@@ -14,39 +14,4 @@ export const txsListener = (socket: Socket, setNewTxs: Function): void => {
       setNewTxs(unconfirmedTxs);
     }
   });
-};
-
-interface ShiftListenerParams {
-  addressType: string
-  socket: Socket
-  setCoins: Function
-  setCoinPair: Function
-  setLoadingPair: Function
-  setSideshiftShift: Function
-  setLoadingShift: Function
-  setSideshiftError: Function
-}
-
-export const sideshiftListener = (params: ShiftListenerParams): void => {
-  params.socket.on('send-sideshift-coins-info', (coins: SideshiftCoin[]) => {
-    params.setCoins(coins.filter(c => c.coin !== params.addressType))
-  })
-  params.socket.on('shift-creation-error', (error: SideshiftError) => {
-    params.setSideshiftError(error)
-    params.setLoadingShift(false)
-    return
-  });
-  params.socket.on('quote-creation-error', (error: SideshiftError) => {
-    params.setSideshiftError(error)
-    params.setLoadingShift(false)
-    return
-  });
-  params.socket.on('shift-created', (shift: SideshiftShift) => {
-    params.setSideshiftShift(shift)
-    params.setLoadingShift(false)
-  });
-  params.socket.on('send-sideshift-rate', (pair: SideshiftPair) => {
-    params.setCoinPair(pair)
-    params.setLoadingPair(false)
-  })
 };
